refactor: moved badges
