+++ quizeek/src/components/quiz/detail/quiz-detail.tsx
@@ -1,10 +1,10 @@
 import { auth } from '@/auth';
 import { QuizDetail as QuizDetailType } from '@/db/schema/quiz';
 
-import QuizAuthorBadge from '../quiz-author-badge';
+import QuizAuthorBadge from '../badge/quiz-author-badge';
+import QuizScoreBadge from '../badge/quiz-score-badge';
+import QuizTimeBadge from '../badge/quiz-time-badge';
 import { QuizImage } from '../quiz-image';
-import QuizScoreBadge from '../quiz-score-badge';
-import QuizTimeBadge from '../quiz-time-badge';
 import { QuizDetailActions } from './quiz-detail-actions';
 import { QuizDetailMyAttemps } from './quiz-detail-my-attempts';
 import { QuizDetailOtherAttemps } from './quiz-detail-other-attempts';
